// src/app/shared/client/clase-cliente.ts
/*
 * To change this license header, choose License Headers in Project Properties.
 * To change this template file, choose Tools | Templates
 * and open the template in the editor.
 * created by oscarcisneros oscarcisneros8@gmail.com
 */
export class ClaseCliente {
  codigo: string;
  descripcion: string;
}